requirements.txt
requests
python-dotenv
pydantic
mcp[cli]